% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_maps.R
\name{available_maps}
\alias{available_maps}
\title{Retrieve available maps in geodato}
\usage{
available_maps(type = "all")
}
\arguments{
\item{type}{A character string specifying the type of maps to retrieve. Default is "all".
Acceptable values are "all", "main", and "regions".}
}
\value{
A character vector of available map names based on the specified type.
}
\description{
This function returns a list of available maps in the geodato package based on
the specified type. The type parameter can be set to "all", "main", or "regions"
to filter the list of available maps.
}
\examples{
available_maps(type = "all")
available_maps(type = "main")
available_maps(type = "regions")
}
