% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topojson.R
\name{centroids_from_topojson}
\alias{centroids_from_topojson}
\title{Calculate centroids from TopoJSON data}
\usage{
centroids_from_topojson(tj)
}
\arguments{
\item{tj}{A TopoJSON object or a file path to a TopoJSON file.}
}
\value{
A data frame containing the centroids for each polygon in the input TopoJSON data.
}
\description{
This function takes a TopoJSON object and calculates the centroids for each polygon.
It can handle both TopoJSON files and 'sf' objects.
}
\examples{
centroids_from_topojson("data-raw/geodato/col/col_bog_localidades/bog-localidades.topojson")
}
