% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topojson.R
\name{change_topojson_ids}
\alias{change_topojson_ids}
\title{Change TopoJSON IDs based on provided codes}
\usage{
change_topojson_ids(tj, codes, write_path = NULL)
}
\arguments{
\item{tj}{A TopoJSON object or a file path to a TopoJSON file.}

\item{codes}{A data frame containing codes with 'id' and 'name' columns to update the TopoJSON IDs.}

\item{write_path}{(Optional) A file path to write the modified TopoJSON object. Default is NULL.}
}
\value{
A TopoJSON object with updated IDs based on the provided codes.
}
\description{
This function takes a TopoJSON object and a codes data frame to change the IDs
in the TopoJSON data based on the provided codes. The function can handle both
TopoJSON files and 'sf' objects. Optionally, the modified TopoJSON object can be
written to a file.
}
\examples{
tj <- gd_tj("col_municipalities")
codes <- data.frame(name = c("NECOCLI", "BARRANQUILLA"), id = c("5490", "8001"))
modified_tj <- change_topojson_names(tj, codes)
}
\keyword{internal}
