% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topojson.R
\name{change_topojson_names}
\alias{change_topojson_names}
\title{Change TopoJSON names with new names}
\usage{
change_topojson_names(tj, new_names, write_path = NULL)
}
\arguments{
\item{tj}{A TopoJSON object or a file path to a TopoJSON file.}

\item{new_names}{A data frame containing the original names in the 'name' column and the new names in the 'new_name' column.}

\item{write_path}{An optional file path where the modified TopoJSON will be saved. Default is NULL (no file will be saved).}
}
\value{
The modified TopoJSON object with updated names.
}
\description{
This function takes a TopoJSON object and a new_names data frame to replace the
original names in the TopoJSON data. If a write_path is provided, the modified
TopoJSON will be saved to the specified file path.
The function can handle both TopoJSON files and 'sf' objects.
}
\examples{
tj <- gd_tj("col_municipalities")
new_names_data_frame <- data.frame(name = c("BOGOTA D.C.", "SAN ANDRES"), new_name = c("Bogotá", "Archipielago de San Andrés"))
modified_tj <- change_topojson_names(tj, new_names_data_frame)
}
