% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topojson.R
\name{check_ids_in_topojson}
\alias{check_ids_in_topojson}
\title{Check if all IDs in TopoJSON data match the provided codes}
\usage{
check_ids_in_topojson(tj, codes)
}
\arguments{
\item{tj}{A TopoJSON object or a file path to a TopoJSON file.}

\item{codes}{A data frame containing codes with an 'id' column to compare with the TopoJSON IDs.}
}
\value{
NULL if all IDs match, otherwise an error message is displayed with the list of mismatched IDs.
}
\description{
This function takes a TopoJSON object and a codes data frame to check if all IDs in the TopoJSON data
match the provided codes. If not, an error message is returned with a list of mismatched IDs.
The function can handle both TopoJSON files and 'sf' objects.
}
\examples{
tj <- gd_tj("col_departments")
codes <-  centroids_from_topojson(tj)
check_ids_in_topojson(tj, codes[1:5,])
check_ids_in_topojson(tj, codes[1:5,]) # error with ids missings
}
