% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_altids}
\alias{gd_altids}
\title{Retrieve alternative IDs for a map in geodato}
\usage{
gd_altids(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve alternative IDs.}
}
\value{
A data frame containing alternative IDs for the specified map.
}
\description{
This function takes a map name and returns its alternative IDs available in the geodato package.
The function also handles region maps by retrieving the corresponding main map alternative IDs.
}
\examples{
gd_altids("col_departments")
gd_altids("world_countries")
gd_altids("world_countries_northern_europe")
}
