% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_altnames}
\alias{gd_altnames}
\title{Retrieve alternative names for a map in geodato}
\usage{
gd_altnames(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve alternative names.}
}
\value{
A data frame containing alternative names for the specified map.
}
\description{
This function takes a map name and returns its alternative names available in the geodato package.
The function also handles region maps by retrieving the corresponding main map alternative names and
filtering them based on the region.
}
\examples{
gd_altnames("col_departments")
gd_altnames("col_municipalities_suroccidente_de_cundinamarca") #this as region without regions
}
