% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_centroids}
\alias{gd_centroids}
\title{Retrieve the centroids data for a map in geodato}
\usage{
gd_centroids(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve the centroids data.}
}
\value{
A data frame containing the centroids data for the specified map.
}
\description{
This function takes a map name and returns its centroids data available in the
 geodato package.
The function also handles region maps by retrieving the corresponding main map
 centroids data.
}
\examples{
gd_centroids("col_departments")
gd_centroids("col_departments_pacifico")
}
