% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_codes}
\alias{gd_codes}
\title{Retrieve geocodes for a map in geodato}
\usage{
gd_codes(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve the geocodes.}
}
\value{
A data frame containing geocodes for the specified map.
}
\description{
This function takes a map name and returns its geocodes available in the geodato package.
The function also handles region maps by retrieving the corresponding main map geocodes and
filtering them based on the region.
}
\examples{
gd_codes("col_municipalities")
gd_codes("world_countries")
}
