% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_paths.R
\name{gd_dir}
\alias{gd_dir}
\title{List directories in a specified path with the option to filter by map name}
\usage{
gd_dir(dir = "data-raw/geodato", regex = "/.*/.*/", map_name = NULL)
}
\arguments{
\item{dir}{The directory path to list. Default is "data-raw/geodato".}

\item{regex}{The regular expression pattern to filter the directories. Default is "/.*/.*/", which matches any directory with two or more levels.}

\item{map_name}{An optional string to further filter the directories by map name.}
}
\value{
A character vector of directories that match the given parameters.
}
\description{
This function uses `fs::dir_ls` to list all directories in a specified path, recursively.
The directories can be filtered by regular expression and/or a map name.
}
\examples{
gd_dir()
gd_dir(map_name = "col_departments")

}
