% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_id_format}
\alias{gd_id_format}
\title{Retrieve the ID format of a map in geodato}
\usage{
gd_id_format(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve the ID format.}
}
\value{
A character string containing the ID format for the specified map.
}
\description{
This function takes a map name and returns its ID format in the geodato package.
The function also handles region maps by retrieving the corresponding main map ID format.
}
\examples{
gd_id_format("col_municipalities")
gd_id_format("col_municipalities_sumapaz")
}
