% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_level}
\alias{gd_level}
\title{Retrieve the level of a map in geodato}
\usage{
gd_level(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve the level.}
}
\value{
A character string containing the level for the specified map.
}
\description{
This function takes a map name and returns its level in the geodato package.
The function also handles region maps by retrieving the corresponding main map level.
}
\examples{
gd_level("col_departments")
gd_level("col_departments_pacifico")
}
