% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_match.R
\name{gd_match_codes}
\alias{gd_match_codes}
\title{Match Geodato Data to a Geodato Map by Codes}
\usage{
gd_match_codes(d, map_name = NULL, col = NULL)
}
\arguments{
\item{d}{The input data frame to be matched.}

\item{map_name}{The name of the geodato map. Defaults to NULL.}

\item{col}{The column name or index to be used for matching by codes. Defaults to NULL.}
}
\value{
A data frame with the matched data based on codes.
}
\description{
This function matches input data to a geodato map based on the provided map_name and the specified code column.
}
\examples{
data <- data.frame(id_depto = c("05", "08", "81", "11"), value = runif(4))
gd_match_codes(data, map_name = "col_departments")
gd_match_codes(data, map_name = "col_departments", "id_depto")
}
