% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_match.R
\name{gd_match_names2}
\alias{gd_match_names2}
\title{Match Geodato Data to a Geodato Map by Names with Two Columns}
\usage{
gd_match_names2(d, map_name = NULL, col = NULL)
}
\arguments{
\item{d}{The input data frame to be matched.}

\item{map_name}{The name of the geodato map. Defaults to NULL.}

\item{col}{A vector of column names or indices to be used for matching by names. Must have a length of 2.}
}
\value{
A data frame with the matched data based on names.
}
\description{
This function matches input data to a geodato map based on the provided map_name and two specified name columns.
}
\examples{
data <- data.frame(id_depto = c("Antioquia", "Antioquia", "Quindio", "Quindio", "Quindio"), muni = c("ALEJANDRIA", "CHIGORODO", "ARMENIA", "BUENAVISTA", "CALARCA"), valor = runif(5))
gd_match_names2(data, map_name = "col_municipalities", col = c("muni", "id_depto"))
}
