% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_meta}
\alias{gd_meta}
\title{Retrieve metadata for a map in geodato}
\usage{
gd_meta(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve the metadata.}
}
\value{
A list containing metadata for the specified map.
}
\description{
This function takes a map name and returns its metadata available in the geodato package.
The function also handles region maps by retrieving the corresponding main map metadata.
}
\examples{
gd_meta("col_departments")
gd_meta("mex_cdmx_localities")
}
