% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_match.R
\name{gd_no_match}
\alias{gd_no_match}
\title{Get Rows with No Match in Geodato Data}
\usage{
gd_no_match(d, map_name, col = NULL)
}
\arguments{
\item{d}{The input data frame to be checked for non-matching rows.}

\item{map_name}{The name of the geodato map. Defaults to NULL.}

\item{col}{A column name or index to be used for matching. Defaults to NULL.}
}
\value{
A data frame containing rows with no match in the geodato data.
}
\description{
This function returns the rows in the input data frame that do not match the
 specified geodato map.
}
\examples{
data <- data.frame(geo_name = c("AR", "COL", "PER", "EC", "BLA"), val = runif(5))
gd_no_match(data, map_name = "world_countries", col = "geo_name")
}
