% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_parent_map_name}
\alias{gd_parent_map_name}
\title{Retrieve the parent map name for a map in geodato}
\usage{
gd_parent_map_name(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve the parent map name.}
}
\value{
A character string containing the parent map name for the specified map, or NULL if no parent map is available.
}
\description{
This function takes a map name and returns its parent map name if available in
the geodato package. The function also handles region maps by retrieving the
corresponding main map parent map name.
}
\examples{
gd_parent_map_name("world_countries")
gd_parent_map_name("world_countries_northern_europe") # agregar
}
