% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_possiblenames}
\alias{gd_possiblenames}
\title{Retrieve possible names for a map in geodato}
\usage{
gd_possiblenames(map_name, with_parent = TRUE)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve possible names.}

\item{with_parent}{A logical value indicating whether to include parent map names in the result. Default is TRUE.}
}
\value{
A character vector containing unique possible names for the specified map.
}
\description{
This function takes a map name and returns its possible names available in the
geodato package, including main names, alternative names, and optionally parent map names.
}
\examples{
gd_possiblenames("world_countries")
gd_possiblenames("world_countries", with_parent = FALSE)
gd_possiblenames("world_countries_western_africa")
gd_possiblenames("col_departments")
length(gd_possiblenames("col_municipalities")) # return names with his parent (deparments)
length(gd_possiblenames("col_municipalities", with_parent = FALSE))
}
