% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_region_codes}
\alias{gd_region_codes}
\title{Retrieve region codes for a main map in geodato}
\usage{
gd_region_codes(map_name)
}
\arguments{
\item{map_name}{A character string specifying the main map name to retrieve region codes.}
}
\value{
A data frame containing region codes for the specified main map, or an error message if the input map is a region map.
}
\description{
This function takes a main map name and returns its region codes available in the geodato package.
If the input map is a region map, the function stops with an error message.
}
\examples{
gd_region_codes("col_municipalities")
gd_region_codes("col_municipalities_alto_putumayo") # This will result in an error
}
