% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_regions}
\alias{gd_regions}
\title{Retrieve regions for a main map in geodato}
\usage{
gd_regions(map_name, path = NULL)
}
\arguments{
\item{map_name}{A character string specifying the main map name to retrieve regions.}
}
\value{
A data frame containing regions for the specified main map, or an error message if the input map is a region map.
}
\description{
This function takes a main map name and returns its regions available in the geodato package.
If the input map is a region map, the function stops with an error message.
}
\examples{
gd_regions("col_municipalities")
gd_regions("col_municipalities_sarare") # This will result in an error
}
