% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topojson.R
\name{gd_simplify}
\alias{gd_simplify}
\title{Simplify a TopoJSON object}
\usage{
gd_simplify(tj, output_size = 3e+05)
}
\arguments{
\item{tj}{A TopoJSON object to be simplified.}

\item{output_size}{The desired output size (in bytes) of the simplified TopoJSON object. Default is 3e5.}
}
\value{
A simplified TopoJSON object.
}
\description{
This function simplifies a TopoJSON object to a specified output size while preserving the original shape.
}
\examples{
tj <- gd_tj("col_municipalities")
simplified_topojson <- gd_simplify(tj, output_size = 0.10)
}
