% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_meta.R
\name{gd_tj}
\alias{gd_tj}
\title{Retrieve the TopoJSON data for a map in geodato}
\usage{
gd_tj(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to retrieve the TopoJSON data.}
}
\value{
A data frame containing the TopoJSON data for the specified map.
}
\description{
This function takes a map name and returns its TopoJSON data available in the geodato package.
The function also handles region maps by retrieving the corresponding main map TopoJSON data and
filtering by the region codes.
}
\examples{
gd_tj("world_countries")
gd_tj("mex_cdmx_colonies")
}
