% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_paths.R
\name{gd_topo_lines}
\alias{gd_topo_lines}
\title{Read topojson file as lines}
\usage{
gd_topo_lines(map_name, system = TRUE)
}
\arguments{
\item{map_name}{The name of the map whose topojson file is to be read.}

\item{system}{Logical, if TRUE (default), the function will look for an absolute system path. If FALSE, a relative path is used.}
}
\value{
A character vector where each element represents a line in the topojson file.
}
\description{
This function reads a topojson file associated with a given map name and returns its content as lines.
}
\examples{
gd_topo_lines(map_name = "col_departments", system = TRUE)
gd_topo_lines(map_name = "col_departments", system = FALSE)

}
