% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_geo.R
\name{is_code_or_name}
\alias{is_code_or_name}
\title{Check if input values are codes or names for a given map}
\usage{
is_code_or_name(v, map_name)
}
\arguments{
\item{v}{A vector of values to be checked.}

\item{map_name}{The name of the map for which the values should be checked.}
}
\value{
A character string: "code" if the majority of the input values are codes, "name" otherwise.
}
\description{
This function checks if the input values are codes or names for the specified map.
}
\examples{
is_code_or_name(c("Bogotá", "Quindio", "Caldas", "Atlantico"), "col_departments")
}
