% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_maps.R
\name{main_or_region_map}
\alias{main_or_region_map}
\title{Determine if a map is a main or region map}
\usage{
main_or_region_map(map_nm)
}
\arguments{
\item{map_nm}{A character string specifying the map name to retrieve the type for.}
}
\value{
A character string representing the type of the map, either "main" or "region".
}
\description{
This function takes a map name and returns its type, either "main" or "region",
 based on the information available in the geodato package.
}
\examples{
main_or_region_map("col_departments")
main_or_region_map("col_departments_andina")
}
