% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_maps.R
\name{map_name_from_region_filter_codes}
\alias{map_name_from_region_filter_codes}
\title{Retrieve geocode ids for a region map}
\usage{
map_name_from_region_filter_codes(map_nm)
}
\arguments{
\item{map_nm}{A character string specifying the map name to retrieve geocode ids for.}
}
\value{
A character vector of geocode ids for the specified region map, or NULL if the
 input map is not a region map.
}
\description{
This function takes a map name and returns the geocode ids for that region if
the input map is a region map. If the input map is not a region map,
the function returns NULL.
}
\examples{
map_name_from_region_filter_codes("world_countries_western_africa")
map_name_from_region_filter_codes("col_departments_andina")
map_name_from_region_filter_codes("col_departments") # its not a region
}
