% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_prep_meta.R
\name{read_meta}
\alias{read_meta}
\title{Read Metadata from a Geodato Map}
\usage{
read_meta(path)
}
\arguments{
\item{path}{The path to the geodato map folder containing the metadata file and related files.}
}
\value{
A list containing the validated metadata and additional information.
}
\description{
This function (TEMPORARY TO ADD EXTERNAL MAPS) reads the metadata from a geodato
 map and validates the associated centroids, altnames, altids and parent_map_name.
}
\examples{
read_meta("data-raw/geodato/col/col_bog_localidades")
}
