% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_prep_meta.R
\name{read_meta_path}
\alias{read_meta_path}
\title{Read Metadata and Associated Files from a Geodato Map Path}
\usage{
read_meta_path(path)
}
\arguments{
\item{path}{The path to the geodato map folder containing the metadata file and related files.}
}
\value{
A list containing the validated metadata, associated files, and additional information.
}
\description{
This function reads the metadata from a geodato map folder and validates the
associated centroids, altnames, and altids.
}
\examples{
read_meta_path("data-raw/geodato/col/col_bog_localidades")
}
