% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_maps.R
\name{search_maps}
\alias{search_maps}
\title{Search available maps in geodato}
\usage{
search_maps(q, type = "all")
}
\arguments{
\item{q}{A character string specifying the search query.}

\item{type}{A character string specifying the type of map to search for. Default is "all".
Acceptable values are "all", "main", and "region".}
}
\value{
A character vector of map names that match the search criteria.
}
\description{
This function searches for maps in the geodato package based on a query string
and an optional type parameter. It returns a character vector of map names that
match the search criteria.
}
\examples{
search_maps("colombia", type = "all")
search_maps("cdmx", type = "main")
search_maps("europe", type = "region")
}
