% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_validate.R
\name{valid_geoids}
\alias{valid_geoids}
\title{Validate Geocodes for a given map}
\usage{
valid_geoids(v, map_name)
}
\arguments{
\item{v}{A vector of geocodes to be validated.}

\item{map_name}{The name of the map for which the geocodes should be validated.}
}
\value{
No return value. The function stops with an error message if any of
the provided geocodes do not match the map's geocodes.
}
\description{
This function checks if the input values are valid geocodes for the specified map.
}
\examples{
valid_geoids(c("05", "08", "11"), "col_departments")
valid_geoids(c("ARG", "BRA", "COL"), "world_countries")
valid_geoids(c("123", "B1RA", "4COL"), "world_countries") # out error
}
