% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_validate.R
\name{validate_altids}
\alias{validate_altids}
\title{Validate Alternate IDs}
\usage{
validate_altids(altids, map_name = NULL)
}
\arguments{
\item{altids}{A dataframe containing the alternate IDs to be validated.}

\item{map_name}{(Optional) The name of the map associated with the alternate IDs. This is used for the error message if validation fails.}
}
\value{
No return value. The function stops with an error message if the required column is missing.
}
\description{
This function checks if the input alternate IDs have the required column: id.
}
