% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_validate.R
\name{validate_altnames}
\alias{validate_altnames}
\title{Validate Alternate Names}
\usage{
validate_altnames(altnames, map_name = NULL)
}
\arguments{
\item{altnames}{A dataframe containing the alternate names to be validated.}

\item{map_name}{(Optional) The name of the map associated with the alternate names.
This is used for the error message if validation fails.}
}
\value{
No return value. The function stops with an error message if any of the
required columns are missing or if the column names do not match.
}
\description{
This function checks if the input alternate names have the required columns: id and altname.
It also checks if the column names match the corresponding centroid column names.
}
