% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_validate.R
\name{validate_centroids}
\alias{validate_centroids}
\title{Validate Centroids}
\usage{
validate_centroids(centroids, map_name = NULL)
}
\arguments{
\item{centroids}{A dataframe containing the centroids to be validated.}

\item{map_name}{(Optional) The name of the map associated with the centroids. This is used for the error message if validation fails.}
}
\value{
No return value. The function stops with an error message if any of the required columns are missing.
}
\description{
This function checks if the input centroids have the required columns: id, name, lon, and lat.
}
\examples{
centroids <- gd_centroids("col_departments")
validate_centroids(centroids)
}
