% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_maps.R
\name{validate_map_name}
\alias{validate_map_name}
\title{Validate map name}
\usage{
validate_map_name(map_name)
}
\arguments{
\item{map_name}{A character string specifying the map name to validate.}
}
\value{
None. If the map_name is not found, the function stops with an error message.
}
\description{
This function validates the provided map_name by checking if it exists in the
list of available maps in the geodato package. If the map_name is not found,
the function stops with an error message.
}
\examples{
validate_map_name("col_departments")
validate_map_name("colombia")
}
