% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_validate.R
\name{validate_regions}
\alias{validate_regions}
\title{Validate Regions}
\usage{
validate_regions(regions, map_name)
}
\arguments{
\item{regions}{A dataframe containing the regions to be validated.}

\item{map_name}{The name of the map associated with the regions.}
}
\value{
No return value. The function stops with an error message if the required
 columns are missing or if the region codes are not in gd_codes.
}
\description{
This function validates the input regions, checking if they have the required
columns and if all region codes are in gd_codes.
}
\examples{
regions <- gd_regions("col_municipalities")
validate_regions(regions, "col_municipalities")
}
