% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_validate.R
\name{validate_topojson}
\alias{validate_topojson}
\title{Validate TopoJSON}
\usage{
validate_topojson(tj, scope, centroids, map_name = NULL)
}
\arguments{
\item{tj}{A TopoJSON object to be validated.}

\item{scope}{A scope value associated with the TopoJSON object.}

\item{centroids}{A dataframe containing the centroids to be checked against the TopoJSON object.}

\item{map_name}{The name of the map associated with the TopoJSON object (optional).}
}
\value{
No return value. The function stops with an error message if the required columns are missing or if the centroids are not included in the TopoJSON.
}
\description{
This function validates the input TopoJSON file by checking if it has the required
columns and if all centroids are included in the TopoJSON.
}
\examples{
tj <- gd_tj("col_departments")
centroid <- gd_centroids("col_departments")
validate_topojson(tj, centroids)
}
