% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topojson.R
\name{write_topojson}
\alias{write_topojson}
\title{Write a TopoJSON object to a file}
\usage{
write_topojson(tj, path)
}
\arguments{
\item{tj}{A TopoJSON object to be written to a file.}

\item{path}{The file path where the TopoJSON object should be saved.}
}
\description{
This function takes a TopoJSON object and writes it to a file specified by the given path.
}
\examples{
#write_topojson(topojson_object, "path/to/save/topojson_file.topojson")
}
